Validate HDGF v6+ chunk Length before narrowing to int (#1075)

+++ poi-scratchpad/src/main/java/org/apache/poi/hdgf/chunks/ChunkHeader.java
@@ -50,7 +50,11 @@ public abstract class ChunkHeader {
             ch.setType((int) LittleEndian.getUInt(data, offset));
             ch.setId((int) LittleEndian.getUInt(data, offset + 4));
             ch.setUnknown1((int) LittleEndian.getUInt(data, offset + 8));
-            ch.setLength((int) LittleEndian.getUInt(data, offset + 12));
+            // Match the v4/v5 branch below: reject lengths that would silently
+            // truncate when narrowing the uint32 to a signed int, rather than
+            // letting a wrapped value flow into the offset arithmetic in
+            // ChunkFactory.createChunk (offset + getLength() + sizeInBytes).
+            ch.setLength(Math.toIntExact(LittleEndian.getUInt(data, offset + 12)));
             ch.setUnknown2(LittleEndian.getShort(data, offset + 16));
             ch.setUnknown3(LittleEndian.getUByte(data, offset + 18));
